Teensy 4 support! (#864)
* Pre-teensy4 work - with a 600Mhz clock, a 1Mhz clock was giving us a clock divider that overflowed a uint8_t - whoops...

* Some tweaks to chipset definitions to help out the Teensy 4 implementation

* Updating the pintest program w/Teensy 4 defs

* Preliminary Teensy 4 support, including hardware SPI and clockless chipsets - no support for parallel output or DMA'd output yet - also not fully tested for all chipsets on all pins, but smoke tested with some chipsets and pin combinations and logic analyzer in the meantime

* Checking in initial block clockless output - it compiles, but no testing yet, so it shouldn't be hooked up anywhere yet.

* Tweak and fix parallel output - still need  to hook it up to the default addLeds setup

* more kicking

+++ platforms/arm/nrf51/fastspi_arm_nrf51.h
@@ -9,7 +9,7 @@
 // A nop/stub class, mostly to show the SPI methods that are needed/used by the various SPI chipset implementations.  Should
 // be used as a definition for the set of methods that the spi implementation classes should use (since C++ doesn't support the
 // idea of interfaces - it's possible this could be done with virtual classes, need to decide if i want that overhead)
-template <uint8_t _DATA_PIN, uint8_t _CLOCK_PIN, uint8_t _SPI_CLOCK_DIVIDER>
+template <uint8_t _DATA_PIN, uint8_t _CLOCK_PIN, uint32_t _SPI_CLOCK_DIVIDER>
 class NRF51SPIOutput {
 
   struct saveData {
